sdk-separate-interpreters: ReducingInterpreter added
