+++ sdk/shared/src/main/scala/org/ergoplatform/sdk/AppkitProvingInterpreter.scala
@@ -1,31 +1,23 @@
 package org.ergoplatform.sdk
 
-import org.ergoplatform.validation.ValidationRules
-import sigmastate.basics.DLogProtocol.{DLogProverInput, ProveDlog}
-
-import java.util
-import java.util.{Objects, List => JList}
-import sigmastate.basics.{DiffieHellmanTupleProverInput, SigmaProtocolPrivateInput}
-import org.ergoplatform._
-import org.ergoplatform.sdk.JavaHelpers.{TokenColl, UniversalConverter}
-import sigmastate.Values.{ErgoTree, SigmaBoolean}
-
-import scala.util.Try
-import sigmastate.interpreter.Interpreter.{ReductionResult, ScriptEnv}
-import sigmastate.interpreter.{ContextExtension, CostedProverResult, HintsBag, Interpreter, ProverInterpreter}
-import sigmastate.exceptions.CostLimitException
-import sigmastate.serialization.SigmaSerializer
-import sigmastate.utils.{SigmaByteReader, SigmaByteWriter}
 import debox.cfor
-import org.ergoplatform.sdk.Extensions.{CollOps, PairCollOps}
-import org.ergoplatform.sdk.utils.ArithUtils
-import org.ergoplatform.sdk.wallet.protocol.context.{ErgoLikeParameters, ErgoLikeStateContext, TransactionContext}
+import org.ergoplatform._
+import org.ergoplatform.sdk.JavaHelpers.TokenColl
+import org.ergoplatform.sdk.wallet.protocol.context.{ErgoLikeParameters, ErgoLikeStateContext}
 import org.ergoplatform.sdk.wallet.secrets.ExtendedSecretKey
 import scalan.util.Extensions.LongOps
-import scorex.crypto.authds.ADDigest
-import sigmastate.{AvlTreeData, VersionContext}
+import sigmastate.Values.SigmaBoolean
+import sigmastate.VersionContext
+import sigmastate.basics.DLogProtocol.{DLogProverInput, ProveDlog}
+import sigmastate.basics.{DiffieHellmanTupleProverInput, SigmaProtocolPrivateInput}
+import sigmastate.eval.Evaluation.addCostChecked
+import sigmastate.interpreter.Interpreter.ReductionResult
+import sigmastate.interpreter.{ContextExtension, CostedProverResult, HintsBag, ProverInterpreter}
+import sigmastate.serialization.SigmaSerializer
+import sigmastate.utils.{SigmaByteReader, SigmaByteWriter}
 
 import scala.collection.mutable
+import scala.util.Try
 
 /**
  * A class which holds secrets and can sign transactions (aka generate proofs).
@@ -39,10 +31,7 @@ class AppkitProvingInterpreter(
       val dLogInputs: IndexedSeq[DLogProverInput],
       val dhtInputs: IndexedSeq[DiffieHellmanTupleProverInput],
       params: ErgoLikeParameters)
-  extends ErgoLikeInterpreter with ProverInterpreter {
-
-  override type CTX = ErgoLikeContext
-  import org.ergoplatform.sdk.Iso._
+  extends ReducingInterpreter(params) with ProverInterpreter {
 
   val secrets: Seq[SigmaProtocolPrivateInput[_, _]] = {
     val dlogs: IndexedSeq[DLogProverInput] = secretKeys.map(_.privateInput)
@@ -52,13 +41,6 @@ class AppkitProvingInterpreter(
   val pubKeys: Seq[ProveDlog] = secrets
     .filter { case _: DLogProverInput => true case _ => false}
     .map(_.asInstanceOf[DLogProverInput].publicImage)
-
-  def addCostLimited(currentCost: Long, delta: Long, limit: Long, msg: => String): Long = {
-    val newCost = java7.compat.Math.addExact(currentCost, delta)
-    if (newCost > limit)
-      throw new Exception(s"Cost of transaction $newCost exceeds limit $limit: $msg")
-    newCost
-  }
 
   /** Reduces and signs the given transaction.
    *
@@ -81,127 +63,10 @@ class AppkitProvingInterpreter(
     var currentCost: Long = baseCost
 
     val (reducedTx, txCost) = reduceTransaction(unsignedTx, boxesToSpend, dataBoxes, stateContext, baseCost, tokensToBurn)
-    currentCost = addCostLimited(currentCost, txCost, maxCost, msg = reducedTx.toString())
+    currentCost = addCostChecked(currentCost, txCost, maxCost, msgSuffix = reducedTx.toString())
 
     val (signedTx, cost) = signReduced(reducedTx, currentCost.toInt)
     (signedTx, txCost + cost)
-  }
-
-  /** Reduce inputs of the given unsigned transaction to provable sigma propositions using
-   * the given context. See [[ReducedErgoLikeTransaction]] for details.
-   *
-   * @note requires `unsignedTx` and `boxesToSpend` have the same boxIds in the same order.
-   * @param boxesToSpend input boxes of the transaction
-   * @param dataBoxes    data inputs of the transaction
-   * @param stateContext state context of the blockchain in which the transaction should be signed
-   * @param baseCost the cost accumulated so far and before this operation
-   * @param tokensToBurn requested tokens to be burnt in the transaction, if empty no burning allowed
-   * @return a new reduced transaction with all inputs reduced and the cost of this transaction
-   *         The returned cost doesn't include (so they need to be added back to get the total cost):
-   *         1) `baseCost`
-   *         2) reduction cost for each input.
-   */
-  def reduceTransaction(
-        unsignedTx: UnsignedErgoLikeTransaction,
-        boxesToSpend: IndexedSeq[ExtendedInputBox],
-        dataBoxes: IndexedSeq[ErgoBox],
-        stateContext: ErgoLikeStateContext,
-        baseCost: Int,
-        tokensToBurn: IndexedSeq[ErgoToken]): (ReducedErgoLikeTransaction, Int) = {
-    if (unsignedTx.inputs.length != boxesToSpend.length) throw new Exception("Not enough boxes to spend")
-    if (unsignedTx.dataInputs.length != dataBoxes.length) throw new Exception("Not enough data boxes")
-
-    val inputTokens = boxesToSpend.flatMap(_.box.additionalTokens.toArray)
-    val outputTokens = unsignedTx.outputCandidates.flatMap(_.additionalTokens.toArray)
-    val tokenDiff = JavaHelpers.subtractTokens(outputTokens, inputTokens)
-    if (tokenDiff.nonEmpty) {
-      val (toBurn, toMint) = tokenDiff.partition(_._2 < 0)  // those with negative diff are to be burnt
-      if (toBurn.nonEmpty) {
-        if (!tokensToBurn.isEmpty) {
-          val requestedToBurn = isoTokensListToTokenColl.to(tokensToBurn.convertTo[JList[ErgoToken]])
-          val diff = JavaHelpers.subtractTokenColls(
-            reducedTokens = toBurn.mapSecond(v => -v), // make positive amounts
-            subtractedTokens = requestedToBurn
-          )
-          if (diff.nonEmpty) {  // empty diff would mean equality
-            throw TokenBalanceException(
-              "Transaction tries to burn tokens, but not how it was requested", diff)
-          }
-        } else {
-            throw TokenBalanceException(
-              "Transaction tries to burn tokens when no burning was requested", tokenDiff)
-        }
-      }
-      if (toMint.nonEmpty) {
-        if (toMint.length > 1) {
-          throw TokenBalanceException("Only one token can be minted in a transaction", toMint)
-        }
-        val isCorrectMintedTokenId = Objects.deepEquals(toMint(0)._1.toArray, boxesToSpend.head.box.id)
-        if (!isCorrectMintedTokenId) {
-          throw TokenBalanceException("Cannot mint a token with invalid id", toMint)
-        }
-      }
-    }
-
-    // Cost of transaction initialization: we should read and parse all inputs and data inputs,
-    // and also iterate through all outputs to check rules
-    val initialCost = ArithUtils.addExact(
-      1000,
-      java7.compat.Math.multiplyExact(boxesToSpend.size, params.inputCost),
-      java7.compat.Math.multiplyExact(dataBoxes.size, params.dataInputCost),
-      java7.compat.Math.multiplyExact(unsignedTx.outputCandidates.size, params.outputCost)
-    )
-    val maxCost = params.maxBlockCost
-    val startCost = addCostLimited(baseCost, initialCost, maxCost, msg = unsignedTx.toString())
-
-    val transactionContext = TransactionContext(boxesToSpend.map(_.box), dataBoxes, unsignedTx)
-
-    val (outAssets, outAssetsNum) = JavaHelpers.extractAssets(unsignedTx.outputCandidates)
-    val (inAssets, inAssetsNum) = JavaHelpers.extractAssets(boxesToSpend.map(_.box))
-
-    val tokenAccessCost = params.tokenAccessCost
-    val totalAssetsAccessCost =
-      java7.compat.Math.addExact(
-        java7.compat.Math.multiplyExact(java7.compat.Math.addExact(outAssetsNum, inAssetsNum), tokenAccessCost),
-        java7.compat.Math.multiplyExact(java7.compat.Math.addExact(inAssets.size, outAssets.size), tokenAccessCost))
-
-    val txCost = addCostLimited(startCost,
-      delta = totalAssetsAccessCost,
-      limit = maxCost, msg = s"when adding assets cost of $totalAssetsAccessCost")
-
-    var currentCost = txCost
-    val reducedInputs = mutable.ArrayBuilder.make[ReducedInputData]
-
-    for ((inputBox, boxIdx) <- boxesToSpend.zipWithIndex) {
-      val unsignedInput = unsignedTx.inputs(boxIdx)
-      require(util.Arrays.equals(unsignedInput.boxId, inputBox.box.id))
-
-      val context = new ErgoLikeContext(
-        AvlTreeData.avlTreeFromDigest(ADDigest @@@ stateContext.previousStateDigest.toArray),
-        stateContext.sigmaLastHeaders,
-        stateContext.sigmaPreHeader,
-        transactionContext.dataBoxes,
-        transactionContext.boxesToSpend,
-        transactionContext.spendingTransaction,
-        boxIdx.toShort,
-        inputBox.extension,
-        ValidationRules.currentSettings,
-        costLimit = maxCost - currentCost,
-        initCost = 0,
-        activatedScriptVersion = (params.blockVersion - 1).toByte
-      )
-
-      val reducedInput = reduce(Interpreter.emptyEnv, inputBox.box.ergoTree, context)
-
-      currentCost = addCostLimited(currentCost,
-        reducedInput.reductionResult.cost, limit = maxCost, msg = inputBox.toString())
-
-      reducedInputs += reducedInput
-    }
-
-    val reducedTx = ReducedErgoLikeTransaction(unsignedTx, reducedInputs.result())
-    val txReductionCost = txCost.toInt - baseCost
-    (reducedTx, txReductionCost)
   }
 
   /** Signs the given transaction (i.e. providing spending proofs) for each input so that
@@ -228,7 +93,7 @@ class AppkitProvingInterpreter(
       val proverResult = proveReduced(reducedInput, unsignedTx.messageToSign)
       val signedInput = Input(unsignedInput.boxId, proverResult)
 
-      currentCost = addCostLimited(currentCost, proverResult.cost, maxCost, msg = signedInput.toString())
+      currentCost = addCostChecked(currentCost, proverResult.cost, maxCost, msgSuffix = signedInput.toString())
 
       provedInputs += signedInput
     }
@@ -237,29 +102,6 @@ class AppkitProvingInterpreter(
       provedInputs.result(), unsignedTx.dataInputs, unsignedTx.outputCandidates)
     val txCost = currentCost.toInt - baseCost
     (signedTx, txCost)
-  }
-
-  // TODO pull this method up to the base class and reuse in `prove`
-  /** Reduces the given ErgoTree in the given context to the sigma proposition.
-   *
-   * @param env      script environment (use Interpreter.emptyEnv as default)
-   * @param ergoTree input ErgoTree expression to reduce
-   * @param context  context used in reduction
-   * @return data object containing enough data to sign a transaction without Context.
-   */
-  def reduce(env: ScriptEnv,
-            ergoTree: ErgoTree,
-            context: CTX): ReducedInputData = {
-    val initCost = ergoTree.complexity + context.initCost
-    val remainingLimit = context.costLimit - initCost
-    if (remainingLimit <= 0)
-      throw new CostLimitException(initCost,
-        s"Estimated execution cost $initCost exceeds the limit ${context.costLimit}", None)
-
-    val ctxUpdInitCost = context.withInitCost(initCost).asInstanceOf[CTX]
-
-    val res = fullReduction(ergoTree, ctxUpdInitCost, env)
-    ReducedInputData(res, ctxUpdInitCost.extension)
   }
 
   // TODO pull this method up to the base class and reuse in `prove`
